refactor: extract scenario projection types
