+++ devtools/quality_registry.py
@@ -2,8 +2,7 @@
 
 from __future__ import annotations
 
-from dataclasses import asdict, dataclass
-from enum import Enum
+from dataclasses import dataclass
 
 from devtools.benchmark_campaign import CAMPAIGNS as BENCHMARK_CAMPAIGNS
 from devtools.benchmark_campaigns import SYNTHETIC_BENCHMARK_SCENARIOS
@@ -12,7 +11,11 @@ from devtools.validation_lane_base import LaneConfig
 from devtools.validation_lane_catalog_composites import COMPOSITE_LANES
 from devtools.validation_lane_catalog_contracts import CONTRACT_LANES
 from devtools.validation_lane_catalog_live import LIVE_LANES
-from polylogue.scenarios import ScenarioMetadata
+from polylogue.scenarios import (
+    ScenarioMetadata,
+    ScenarioProjectionEntry,
+    ScenarioProjectionSourceKind,
+)
 from polylogue.showcase.exercises import EXERCISES
 
 
@@ -42,45 +45,6 @@ class BenchmarkCampaignEntry(ScenarioMetadata):
     notes: tuple[str, ...] = ()
     warn_pct: float = 0.0
     fail_pct: float = 0.0
-
-
-class ScenarioProjectionSourceKind(str, Enum):
-    EXERCISE = "exercise"
-    BENCHMARK_CAMPAIGN = "benchmark-campaign"
-    SYNTHETIC_BENCHMARK = "synthetic-benchmark"
-
-
-@dataclass(frozen=True)
-class ScenarioProjectionEntry(ScenarioMetadata):
-    source_kind: ScenarioProjectionSourceKind
-    name: str
-    description: str
-
-    @classmethod
-    def from_object(
-        cls,
-        *,
-        source_kind: ScenarioProjectionSourceKind,
-        name: str,
-        description: str,
-        obj: object,
-    ) -> ScenarioProjectionEntry:
-        metadata = ScenarioMetadata.from_object(obj)
-        return cls(
-            source_kind=source_kind,
-            name=name,
-            description=description,
-            origin=metadata.origin,
-            path_targets=metadata.path_targets,
-            artifact_targets=metadata.artifact_targets,
-            operation_targets=metadata.operation_targets,
-            tags=metadata.tags,
-        )
-
-    def to_dict(self) -> dict[str, object]:
-        data = asdict(self)
-        data["source_kind"] = self.source_kind.value
-        return data
 
 
 @dataclass(frozen=True)
