+++ polylogue/scenarios/__init__.py
@@ -7,9 +7,15 @@ from .metadata import (
     runtime_operation_target_names,
     runtime_path_target_names,
 )
+from .projections import (
+    ScenarioProjectionEntry,
+    ScenarioProjectionSourceKind,
+)
 
 __all__ = [
     "ScenarioMetadata",
+    "ScenarioProjectionEntry",
+    "ScenarioProjectionSourceKind",
     "runtime_artifact_graph",
     "runtime_artifact_target_names",
     "runtime_operation_target_names",
